LayoutTests/platform/chromium/accessibility/footer-expected.txt

PASS accessibilityController.focusedElement.role is "AXRole: AXGroup"
PASS accessibilityController.focusedElement.roleDescription is "AXRoleDescription: footer"
PASS accessibilityController.focusedElement.role is "AXRole: AXGroup"
PASS accessibilityController.focusedElement.roleDescription is "AXRoleDescription: group"
PASS accessibilityController.focusedElement.role is "AXRole: AXGroup"
PASS accessibilityController.focusedElement.roleDescription is "AXRoleDescription: group"
PASS successfullyParsed is true

TEST COMPLETE
Footer outside section and article elements.
Footer inside an article.
Footer inside a section.
